moved plugin files outside of root
