+++ module.go
@@ -6,6 +6,8 @@ import (
 	"github.com/caddyserver/caddy/v2/caddyconfig/httpcaddyfile"
 	"github.com/caddyserver/caddy/v2/modules/caddyhttp"
 	"github.com/caddyserver/caddy/v2/modules/caddyhttp/reverseproxy"
+	mod "github.com/openrelayxyz/din-caddy-plugins/modules"
+	prom "github.com/openrelayxyz/din-caddy-plugins/services/prometheus"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -14,34 +16,34 @@ var (
 	// https://caddyserver.com/docs/extending-caddy
 
 	// Din Middleware Module
-	_ caddy.Module                = (*DinMiddleware)(nil)
-	_ caddy.Provisioner           = (*DinMiddleware)(nil)
-	_ caddyhttp.MiddlewareHandler = (*DinMiddleware)(nil)
-	_ caddyfile.Unmarshaler       = (*DinMiddleware)(nil)
+	_ caddy.Module                = (*mod.DinMiddleware)(nil)
+	_ caddy.Provisioner           = (*mod.DinMiddleware)(nil)
+	_ caddyhttp.MiddlewareHandler = (*mod.DinMiddleware)(nil)
+	_ caddyfile.Unmarshaler       = (*mod.DinMiddleware)(nil)
 	// TODO: validate provision step
-	// _ caddy.Validator			= (*DinMiddleware)(nil)
+	// _ caddy.Validator			= (*mod.DinMiddleware)(nil)
 
 	// Din Upstream Module
-	_ caddy.Module                = (*DinUpstreams)(nil)
-	_ reverseproxy.UpstreamSource = (*DinUpstreams)(nil)
+	_ caddy.Module                = (*mod.DinUpstreams)(nil)
+	_ reverseproxy.UpstreamSource = (*mod.DinUpstreams)(nil)
 
 	// Din Select Module
-	_ caddy.Module      = (*DinSelect)(nil)
-	_ caddy.Provisioner = (*DinSelect)(nil)
+	_ caddy.Module      = (*mod.DinSelect)(nil)
+	_ caddy.Provisioner = (*mod.DinSelect)(nil)
 )
 
 func init() {
-	caddy.RegisterModule(DinUpstreams{})
-	caddy.RegisterModule(DinSelect{})
-	caddy.RegisterModule(DinMiddleware{})
+	caddy.RegisterModule(mod.DinUpstreams{})
+	caddy.RegisterModule(mod.DinSelect{})
+	caddy.RegisterModule(mod.DinMiddleware{})
 	httpcaddyfile.RegisterHandlerDirective("din", parseCaddyfile)
 
 	// Register custom prometheus request metrics
-	prometheus.MustRegister(dinRequestCount)
+	prometheus.MustRegister(prom.DinRequestCount)
 }
 
 func parseCaddyfile(h httpcaddyfile.Helper) (caddyhttp.MiddlewareHandler, error) {
-	j := new(DinMiddleware)
+	j := new(mod.DinMiddleware)
 	err := j.UnmarshalCaddyfile(h.Dispenser)
 	if err != nil {
 		return nil, err
