+++ modules/din_middleware.go
@@ -1,4 +1,4 @@
-package din
+package modules
 
 import (
 	"fmt"
