+++ services/prometheus/prometheus.go
@@ -1,9 +1,9 @@
-package din
+package prometheus
 
 import "github.com/prometheus/client_golang/prometheus"
 
 // prometheus metric initialization
-var dinRequestCount = prometheus.NewCounterVec(
+var DinRequestCount = prometheus.NewCounterVec(
 	prometheus.CounterOpts{
 		Name: "din_http_request_count",
 		Help: "",
